+++ tests/wctype/towlower.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <stdio.h>
 #include <wchar.h>
 #include <wctype.h>
@@ -5,7 +6,7 @@
 int main() {
     wchar_t *str = L"HaLf WiDe ChAr StRiNg!\n";
     fputws(str, stdout);
-    for (int i = 0; i < wcslen(str); i++) {
+    for (size_t i = 0; i < wcslen(str); i++) {
         putwchar(towctrans(str[i], wctrans("tolower")));
     }
     return 0;
